DC-274: integration tests with snapshots (#53)
Add an integration test for all service operations. The test creates, updates, retrieves and deletes a catalog entry for a TDR snapshot.

While testing we found that the HTTP status code for a missing catalog entry record was 500, but should've been 404. This PR also fixes this behavior and adds a unit test for this case.