Update the swagger-client-id path
Previously, we had client ids which lived at the following paths:

1. `secret/dsde/terra/kernel/dev/dev/catalog/swagger-client-id`
2. `secret/dsde/terra/kernel/dev/dev/catalog`

To be consistent with other services, the client id should live at `{service}/swagger-client-id`, however we had an invalid client id at that path and a valid one at the `/catalog` path. This fixes the path issue in the `render-configs.sh` script.