Add a gradle subproject for scripts. (#69)
Add a gradle subproject for scripts. This migrates the `shellcheck` and `black` operations from GitHub actions to gradle tasks, which makes it easier to run these operations locally.

With this change, `black` and `shellcheck` are triggered via the `check` task, which is run by `test`. So running `build` at the root level will also run `black` and `shellcheck` on the scripts subproject, if that project is out of date.

To use this, you must have `shellcheck` installed at `/usr/local/bin/shellcheck` (which is where `brew install shellcheck` puts it), and `black` must be installed and on your `PATH` (using `pip install black` should do this).

Another nice thing about having a subproject for scripts is it makes it easier to configure IntelliJ to support python editing features for the `.py` files.

This change also adds python settings to `.editorconfig`.