DC-361: Dynamically create TDR dataset once for all integration tests (#89)
To fix issues with contention when running integration tests, the test code now creates a TDR dataset on demand. Only one dataset is created and is used for all tests, and a JVM shutdown hook is used to ensure that test datasets are deleted from TDR when the test script exits.