DC-523: Fix missing content-length issue by adding a body to the POST (#124)
The swagger generated Java client library doesn't include a Content-Length header if the POST body is null (instead of providing Content-Length: 0, which our OIDC proxy server expects). There are ways to fix this by changing the client code behavior, and it may even be a bug in the swagger codegen, but instead of trying to fix the problem there, I think it's better to move the destination workflow ID parameter to a body parameter instead of a path parameter. We believe that the catalog will need to support different export destinations in the future, at which point we'd need to add more flexibility to the API anyway.