DC-278: Use a session scoped bean for the authenticated user (#37)
Based on a suggestion from @petesantos, this change uses a http request scoped bean to generate the current authenticated user's token. The notable improvement of this approach is that the code no longer needs to pass the request's user all the way from the controller down to the service layer. With this change, the `User` is only injected in the components that needs it, which currently are `DatarepoClient`, `RawlsClient` and `SamService`.

This also removes the code that used SAM to retrieve the current user's email. Our code doesn't use the email, and we shouldn't be using SAM for this purpose anyway; in general we shouldn't call SAM on every incoming request unless it's necessary. If we need the user email, we should be extracting it from the OIDC proxy generated headers.