DC-524: Add script to validate metadata in the catalog database (#156)
Implement a validation script for the catalog as a spring command line runner.

This change restructures the code in the repo so that most of the spring services are now in a new subproject, `common`. The `service` subproject now only contains the controller classes and other web API deployment components.  Another new subproject, `admin-cli`, contains the new CLI program.