DC-693: service startup error fix
One theory is that oen of the status checks are taking too long which causes the status check concurrent task to no longer fire. This changes the concurrent task to use `scheduleWithFixedDelay()` instead of `scheduleAtFixedRate()`, to guard against the case where a status check takes longer than the thread interval.

When doing this, I noticed that the timeout for the database check was using 5000 seconds. This check should be very fast so I'm changing the timeout to 1 second instead.

This change also adds unit test coverage for the database status check.