DC-839: Move references back to datarepo-client; Upgrade to latest (#175)
https://broadworkbench.atlassian.net/browse/DC-839
_______
With the https://github.com/DataBiosphere/jade-data-repo/pull/1581 for
data repo, we're switching back to only publishing to one Jakarta backed
client. This PR moves your references to the jakarta client back to the
main data repo client and bumps to the latest version. This is an effort
to alleviate any confusion around upgrading past the last supported
datarepo-jakarta-client version.