[DT-1022] Add CSP to the Swagger UI using meta tag (#182)
### Addresses

https://broadworkbench.atlassian.net/browse/DT-1022

### Summary

Similar to
https://github.com/DataBiosphere/terra-landing-zone-service/pull/502 and
https://github.com/DataBiosphere/terra-workspace-data-service/pull/941 ,
adds a Content Security Policy to the Swagger UI using a `meta` tag.

The team decided to use this approach instead of the proxy approach
since catalog currently has no proxy templating which means that we
would either need to modify it for all services or add our own template
for catalog, both of which were deemed high risk and low value,
especially since catalog is likely to be retired in the near future.
Additionally, proxy changes are hard to test, whereas this change was
straightforward to test.

### Testing

Tested with a catalog instance locally, tried endpoints and the login,
no errors were observed in the console. Next, added a CDN delivered
resource, and saw failures relating to that resource in the console.